6808315 sdpadm can no longer enable sdp on SPARC systems 6800771 sdpib fails to initialize on sparc platforms

+++ usr/src/uts/common/io/ib/clients/sdp/sdpddi.c
@@ -19,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2008 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2009 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -181,7 +181,7 @@ sdp_gen_ioctl(queue_t *q, mblk_t *mp)
 	/* LINTED */
 	iocp = (struct iocblk *)mp->b_rptr;
 	switch (iocp->ioc_cmd) {
-			uintptr_t send_enable;
+			int32_t send_enable;
 		case SIOCSENABLESDP:
 			bcopy(mp->b_cont->b_rptr, &enable, sizeof (int));
